Fix: Bone parenting works
before there was no attempt to trace backward through the tree to find the parent. the parent was requested in Edit mode for some reason. and because the armature was posed, the inverse matrix was all wrong. All of that is fixed, and as a bonus, it should fix automatic weight painting, too.
also fix object parenting, which was checking if the mantis node was a Blender Obejct and trying to set the parent to the node.