Fix Copy Weights in Armature Deformer
fixed the copy-weights feature not initializing the vertex groups, fixed the unnecesary error when there was no object to copy from and replaced it with a warning message, improved display update code.
Note: I may have Armature Deform (Copy Weights) apply the Data Transfer modifier and do some smart stuff about sampling topology  or UVs or something. This is not adressed here and is left for the future.