Fix: Array Inputs now sort correctly
Before this fix, there was no logic to handle the order of the array node's inputs, so they were added in order of the node_tree.links() list (which is ordered by lifetime, newest first, and this is incidentally the same as sorting by the ram/pointer address). Now we look at the sort index stored in the bpy.types.link object, and use that. My implementation is probably not the most efficient but it's fast enough.