Fix: Pole Vector chooses correct bend axis
Now, we find out where the "knee" is in the IK chain and check which of
the base-bone's axes it is in-line with. Then we calculate the poll
vector based around the perpendicular axis.
My simple tests worked perfectly. Need to test this more since the
math is confusing.
Also added some math utilities to utilities.py. Bisect isn't used
but since I went to the trouble to find it and it is useful, I'm
keeping it in, in case I need it later.