Fix: fail to import node groups because interface not set
Previously, I attempted to sort the node trees as I import them based on their depth. There were bugs because the way I was sorting them was not robust, thus I am opting to set up each node group 's interface before trying to process the import. This solves the problem of NodeGroup's with the wrong number of inputs.