Operators tfor making bulk connections
Select All Nodes of Type selects all nodes sharing the same type (and tree, for node groups). ConnectNodeToInput connects a Node Group Input to all sockets in selected nodes if possible.
This is not optimized, and could be sped up by recording and using the indices of the node sockets, but this is safe and much faster than doing it by hand