Fix: Group Node Enter/Exit causes tree execution
despite my precautions, the Edit Group operator caused the tree to execute.
Turns out, it was happening outside of the operator, after it finished executing.
So, now the tree has a new attribute that does nothing but prevent it from executing under these circumstances
did a little bit of safety cleanup stuff with a finally block