Feature: Nodes show names
this is a first hack at making the node graph show xForm's names and the result of Cat String nodes
in future I will be looking to make this sort of interactive feedback more pervasive and reliable.
This commit also fixes some problems with the xFormBone display update code that may have interfered with the previous commit
finally there is a tiny bit of cleanup in the Matrix nodes' code.