Refactor Schema Solve
This commit splits schema solve into more managble functions, removes some
unnecessary variables, renames things to be more readable, and simplifies the
handling of node signatures (although it still isn't perfect).

Further work needs to be done in ensuring the graph is prepared for the schema
solve itself, which may need more than just its length, and should catch and raise
errors if it finds a cycle.

This patch doesn't resolve any bugs as far as I am aware -- it is primarily a refactor.