Optimize: NodeSocket.is_linked now faster
using a property decorator and checking the socket for links was slow because
it added two function calls - once for the property and once for the bool(list).

Now we just set the property to True when we add a link with connect() and False
 if flush_links detects an empty list.

This function was one of the bigger time-users, now it uses 0 time.