Refactor & Optimize: UI classes now have a mantis class reference
This commit ends up saving ~10% time in the component_groups test file. (from ~1.1s to ~0.9s).
The biggest change is the fact that there is no costly lookup function finding mantis classes during
tree parsing. Now, the classes get a class reference assigned at init - even better, if the class reference
is not found, the class won't init and the addon will fail to load. Thus, it is also an effective error-
finding solution for the problem it solves.
The code is a little ugly, but not terribly so, and it provides a stopgap solution while I prepare the
much bigger change of generating UI classes programatically. So the ugliness is temporary.

There is also a tiny change in MantisNodeSocketCollection which removes a custom __setitem__
which, though harmless, was costing me precious milliseconds.

the change in schema_containers is formatting/cleanup.