Fix: Inherit Node sockets will not hide if they are connected.
This is a slapdash solution because these nodes are very common and it will take
a little more work to do this everywhere in a clean way. Basically, sockets that are
connected but should be hidden should not hide, and maybe they should even
have an error color or message. I judge this particular situation to be common
enough to warrant a fix now.