Fix: Empty Matrix Socket in Node Group causes error
This bug was caused because the __init__ of the InputMatrix node expected a ui node
to copy data from - even though it was an autogenerated node from a node group
input. I fixed it by removing that section entirely. I don't know why it was there in the
first place.

tested to ensure the matrix node still functions as expected.