enable blender 4.4.0
this commit hacks around a bug in the Python API regarding NodeTree.valid_socket_get()
and MantisUINode.insert_link(). I expect these to be fixed in a point release soon.
otherwise my addon mostly works out of the box.

This will not work in 4.4 directly, there seems to be some sort of tree parsing problem - but I
suspect it will be helpful to fix this for 4.3 too. so it will be the next commit.