Cleanup: Socket Type menu is now much more managable
Added a class property to the NodeSocket types to mark them as valid interface types or not.
This also lays some groundwork for more cleanup to happen soon.
This doesn't seem to have any harmful effects on group nodes, since the
interface type CAN be set to any socket bl_idname, and it won't clutter
the console or print errors - but the user cannot manually set those
types (which is good).

also added a type annotation since I need to start doing that