Fix: Node Groups don't work properly in Schema
This commit makes it so Node Groups within Schema are treated as schema with
length == 1
now I am not 100% sure this is going to work in every circumstance but it should
work anywhere schema does

this commit does NOT guarantee that the parsed tree is functional - it does make
sure the tree will parse properly. So there may still be bugs in some nodes and I
will have to fix these on a case-by-case basis.