New Socket Template class and methods
This commit introduces the dataclass and methods for the unified socket generation goal,
which will make it so I can declare the sockets in one place and then use them to initialize both
the node-container and the ui-node
The result of this will be simpler code, less total lines of code, reduction of both boilerplate and
code duplication. Thus it will also lead to stability improvements.
Finally it also means that I can set default values and make copy constructors.
This commit may also have beneficial ramifications for versioning, too, since it could be used
to generate nodes again from scratch very easily.