fix: node update broken due
this one was because I was trying to use dict-getters on node sockets.
I have no idea why I ever thought that was a good idea, it was
probably just an accident
anyhow I changed it to getattr.
I don't know how this section of code ever worked. I'm amazed that it did