Fix: Curves getting wiped from Node Group inputs
I honestly don't know what was causing this, but this fixes it by
making the conditions for the recent (very small) changes I made
much more specific.
I need the new behaviour that sort of broke things, but I only need it
when  conditions require me to reinitialize the node.