Fix: Node Tree Update only updates the sockets it found when starting update
Basically what this commit does, it prevents the update function from clearing any socket that can be reused - or should be reused because there was an error.
It also makes it so that sockets which aren't found but are declared by the interface
will be added, but there will be no attempt to run do_relink() on them.