Schema Solve: prepare missing dependencies
This is a bad way to fix what is probably an error in graph generation.
some nodes need to have their parents prepared, and the parents aren't in the graph
we solve this by additionally preparing nodes that have children in the solve_only_these
list.
I ran into this situation when adding the new Array Length and Choose Nearest nodes.
Probably I can fix this in another way, and should keep trying because I want
the node execution/parsing code to be as simple as possible.