Add: GetCurvePoint, GetNearestFactorOnCurve
this commit uses some old code from Chaintools, almost unedited.
in FindNearestPointOnWireMesh, I'm not 100% sure I am calculating the distance correctly
(what's the deal with the sqrt's when I am already getting the length of the vector?)
so I may have to revisit it. but it seems to work real well.