cleanup Curve Pin & add locks
during this commit's development, I experimented with setting a parent on the curve pin. I decided against it... the curve pin is already sort of a child of
the curve, and setting the matrices is really tricky when parents are involved.
moreover, when i did it manually to test, I found no real benefit to setting the parents
perhaps it was even messing things up, it certainly complicates things.