Update Hook Deformer
hook deformer now has an input for strength, and can automatically set Bezier Handle hooks for easier use (indexing by bezier point instead of NURBS).
the ability to directly hook bezier handles is retained if you turn off the auto-bezier option.

If the strength of two hook deformers sharing a common controller is the same, the modifier
will be reused. But if there is a driver, or the strength is different, it will make a new modifier
TODO: I should check and see if the drivers are the same to reuse the modifer even if
there is a driver. but this isn't very important, so I haven't done it.