Add MantixExecutionContext
it isn't pretty and it increases complexity just a little bit
but I need to be able to store some extra information

I'm caught in a development paradox here. refactoring mantis to include this feature
the "right" way - by adding it to class init - would take a lot of work. Adding it at runtime
is still class composition, which is good, but spreads the code out to two different functions
so the only real danger is that I may overlook something and fail to set the value. which is
likely if I make big changes to how things are solved.

thing is, it simply isn't worth it to go to all this trouble until I get some benefit from the new
MantisExecutionContext.
So I am gonna do it this way - the messy, but easy and simple way - because it will free me to
start using the feature. Then, I will observe how my development is affected before I decide
what to do. After all... there are other ways to solve some of these problems.