Fix: Reset Execution to prevent outdated references
this commit is part of an ongoing effort to streamline execution
perviously, execution state was not reset when trying to re-execute nodes on e.g.
socket value changes
so the result was, pointers in Blender's memory were reused and swapped around
and Mantis still held references to them (in Python). as a result, drivers were being
set on completely wrong properties in completely wrong bones
there is also a fair bit of cleanup here

the solution is to reset the bObject value in each node. I've gone ahead and
reset the prepared, executed, and drivers fields, too.

next task: use hashes to track the state of the tree's topology to
 prevent parsing the tree except when graph topology changes

future work:
 - separate execution passes into Parse, Execute, Link, and Finalize passes
 - ensure that parse-work is not repeated during execution