define socket templates for Bone
this commit does not actually IMPLEMENT socket templates for the node
there are a few obstacles to finishing implementation:
 - testing will be a lot of work
 - bbone implementation is still not finished
 - bone is old code that does a lot of stuff in a weird way
 - bone needs a lot of bespoke functionality instead of inherited methods
     - ui_modify_socket and update_socket_value
     - gen_property_socket_map is inadequate for Bone's needs
 - UI code will require reqriting to use the new category system

ultimately, there isn't a very big benefit to doing this right now.
I intend to move to the new paradigm soon but not yet.
I will commit this now, however, because it doesn't affect mantis' actual behavior just yet