Fix: Hierarchy relationships not set by array outputs
I have no idea why this wasn't a problem in my previous tests!
this is a problem I anticipated and attempted to circumvent in previous work,
but as it didn't seem to happen and calling init_connections and init_dependencies
seemed to break things, I left it as it was.

now that I have connected a Schema array output to an array get, I am seeing this problem
Maybe it existed before? Maybe it has showed up because of some bug fix elsewhere?
whatever the case, this fixes it now.