Fix: Schema Index input erases custom property
this one was caused by an overly aggressive automation that deleted some
node inputs and replaced them with evaluated values. Custom properties,
however, need their sockets. The result was that the custom property was not only
not set correctly, but not created at all!

this commit just makes it check first if it is safe to erase