Fix: Blender freezes when using Array Get
this was happening specifically because of this:
 -A schema generates a number of bones and sends them into an array out
 - the array out goes straight into an Array Get
 - the array Get leads to a Matrix of xForm which depends (because of
      the array-get) on a bone from the schema
now at this point, the bone and array are prepped and ready. the Matrix
of xForm node, however, cannot execute because when it scans its
dependencies, it finds that the bone is not in the 'schema_solve_done'
set, because it was created inside of a schema.

I should look into going back to checking for .prepared == True, but I
don't know if I am using schema_solve_done for some other reason. have
to experiment.
I don't think there are any serious performance problems here, though