cleanup schema
this is mostly a cleanup commit, but it also makes a small change in logic
because I think it was overlooked when I wrote this code a few days ago.
Still, the old code seemed to work. I think this is a safe change to make that
will potentially prevent confusing bugs.