fix: subschema cannot link into parent schema
this happened because the subschema requested a node in
the parent schema which was not in the all_nodes collection yet
because it was a member of the same frame as the
sub-schema. the solution is simple: just keep all_nodes
up-to-date

looking back at the code... were the schema nodes ever added
to all_nodes? it looks like the all_nodes was just all the
nodes that existed before the schema was generated?

this commit also does a little cleanup that is not intended to
change behaviour, but might in some rare cases. If it does, I
expect the changes only to be positive.