fix: link from two identical bones ignored in array
this commit softens a check for duplicate links that causes some
configurations of Schema to fail.
I don't think duplicate links do anything bad in general
and they are 100% valid in case of arrays.

I question the wisdom of silently failing instead of raising an error
when these duplicate links are attempted... but I will leave it how
it is for now.