Fix: Array-Type to Schema Array output fails
the problem:
array re-linking has to be solved to resolve dependencies,
but it needs to be wired at both ends.
however, we do not wire the output until it has been prepared
catch 22
so instead, we solve the array nodes as normal and then
wire the links they rerouted out of the array output

kind of confusing and complicated, and a very rare
edge case. I now desire very badly to refactor the data types
of the links in schema solve

having to deal with ui links, mantis dummy links, and mantis links
all together is terrible