Copy node list when solving schema (consistency)
the sub-schema get a copy, so the parent schema should, too.
this also prevents schema from adding interface/prototype
stuff in the main node collection.

if this leads to bugs, then the sub-schema should get a reference
instead of a copy, too, when I change it back.

I haven't encountered any problems in testing.