Fix: ensure autogen nodes are always unique
this bit of code is used by Schema, where previously held
assumptions about its safety no longer apply...
hence: extremely confusing bug that took >10 hours to find the source of

without this uuid signature and the (mostly redndant) check to see
if the node already exists, it is possible to generate nodes
which rely on the nodes that are now shadowed by new nodes.
thus, these nodes never execute because their dependencies
can never be resolved - since the nodes they depend on are no longer
in the parsed tree.

The pattern the error follows is: within a schema, only the most recent nodes in the schema execute. Older nodes are left behind
and cause the tree to fail without detecting a specific error.