Fix: Schema fails for collection and color nodes.
Caused by the ui_signature being wrong in schema-generated
nodes, because the ui_signature is just the signature
during init for most nodes
so the ui_node was not found for schema-generated nodes.

I was too lazy to put the code that reads the UI node in
the fill_parameters function, and really I shouldn't have to
but I don't want to refactor the classes to accept ui_signature
as a parameter in __init__ so I have to be content with
setting it later, after init, from within the Schema solve

whatever. I should refactor this but I don't want to risk bugs. It is what it is.