UI: Widget Socket handle missing socket
in case of missing widgets, the Widget Socket keeps the data around
additionally: the path is now only saved relative to the root
the widget now considers sub-folders in the name
Finally: this commit introduces an idiot check for the widget folder
to prevent the user from freezing Blender with an absurdly large
file system (e.g. the user's home or documents folder).