Fix: Custom properties fail in Drivers
I think - but I haven't confirmed - that newer Blenders have introduced
a small change in how properties are read in Python.
This commit explicitly checks if the custom property is in the .keys()
of the bpy_struct instead of checking for the attribute. I expected
Python to distinguish between a dot-addressed attribute and a bracket-
addressed attribute (.attr vs. ["attr"]) but I guess not or not anymore.

Anyhow the solution is simple and it seems to work in 4.5 and 4.2
without any problems.