Fix: Nested Choose fails when linked to group output
this bug occurs when:
 - the choose node is connected directly to Group output
 - the Group is inside a Schema

Perhaps i should have opted to add a no-op node so that the choose node can go ahead and prepare. Who knows? maybe I willl choose that solution in the future

this bug fix seems to cause no problems in existing trees,
and it fixes the behaviour in old trees that did not work.