Refactor: xForm_info data class for setting parents
This is a pretty BIG refactor that changes the way xForms set their parents
Now, we send parenting info in the node and push it down the links.
Setting parents is a bit cleaner since we can look up the info directly from the
strings, instead of seeking backward through the tree and getting the info
from the mantis node directly

We still need to do some seeking through the tree, unfortunately, but in time
even that will be refactored out.

This seems to perform equivalently or *maybe* a tiny bit faster than before
 the commit. Finally, I think the code is a little less bad now, even though there
 is some ugly stuff I had to do. Whatever. It's an important step.