Custom Properties Nodes (for all xForms!)
this is a really big commit, but I did test it pretty thoroughly
it adds custom properties to all xForms by way of a new Custom
Properties inputon xForm nodes and a new Custom Property
node which allows the user to declare these custom properties
using nodes. This will make it possible to build e.g. parent switches
with the right property min/max for the number of parents in the
input array.

Next task: custom properties for 5.0