Fix: object constraints stopped working
this one happened because I was trying to program too fast
and I missed a check for something when I was factoring in
a new profiler. Amusingly this was the only part of the code
for that commit that I didn't let the LLM touch. Maybe humans
are the weak link and the machine overlords really should take over